[Apicurio] Regen JAX-RS with support for state changes.
