+++ common/src/main/resources/META-INF/openapi.json
@@ -1481,7 +1481,7 @@
                 },
                 "operationId": "updateArtifactState",
                 "summary": "Update Artifact State",
-                "description": "Updates the state of the artifact.  This can be used to, for example, mark the latest\nversion of an Artifact as `DEPRECATED`.  The operation will change the state of the\nlatest version of the artifact.  If multiple versions exist, only the most recent will\nbe changed.\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* Artifact cannot transition to the given state (HTTP error `400`)\n* A server error occurred (HTTP error `500`)\n"
+                "description": "Updates the state of the artifact.  This can be used to, for example, mark the latest\nversion of an Artifact as `DEPRECATED`.  The operation will change the state of the\nlatest version of the artifact.  If multiple versions exist, only the most recent will\nbe changed.\n\nThe following state changes are supported:\n\n* Enabled -> Disabled\n* Enabled -> Deprecated\n* Enabled -> Deleted\n* Disabled -> Enabled\n* Disabled -> Deleted\n* Disabled -> Deprecated\n* Deprecated -> Deleted\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* Artifact cannot transition to the given state (HTTP error `400`)\n* A server error occurred (HTTP error `500`)\n"
             },
             "parameters": [
                 {
@@ -1527,7 +1527,7 @@
                 },
                 "operationId": "updateArtifactVersionState",
                 "summary": "Update Artifact Version State",
-                "description": "Used to update the state of a specific version of an Artifact.  For example, this can\nbe used to \"disable\" a specific version.\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* No version with this `version` exists (HTTP error `404`)\n* Artifact version cannot transition to the given state (HTTP error `400`)\n* A server error occurred (HTTP error `500`)\n"
+                "description": "Used to update the state of a specific version of an Artifact.  For example, this can\nbe used to \"disable\" a specific version.\n\nThe following state changes are supported:\n\n* Enabled -> Disabled\n* Enabled -> Deprecated\n* Enabled -> Deleted\n* Disabled -> Enabled\n* Disabled -> Deleted\n* Disabled -> Deprecated\n* Deprecated -> Deleted\n\nThis operation can fail for the following reasons:\n\n* No artifact with this `artifactId` exists (HTTP error `404`)\n* No version with this `version` exists (HTTP error `404`)\n* Artifact version cannot transition to the given state (HTTP error `400`)\n* A server error occurred (HTTP error `500`)\n"
             },
             "parameters": [
                 {
@@ -1609,14 +1609,8 @@
                         "type": "integer"
                     },
                     "state": {
-                        "description": "",
-                        "enum": [
-                            "ENABLED",
-                            "DISABLED",
-                            "DEPRECATED",
-                            "DELETED"
-                        ],
-                        "type": "string"
+                        "$ref": "#/components/schemas/ArtifactState",
+                        "description": ""
                     }
                 },
                 "example": {
@@ -1683,14 +1677,8 @@
                         "type": "integer"
                     },
                     "state": {
-                        "description": "",
-                        "enum": [
-                            "ENABLED",
-                            "DISABLED",
-                            "DEPRECATED",
-                            "DELETED"
-                        ],
-                        "type": "string"
+                        "$ref": "#/components/schemas/ArtifactState",
+                        "description": ""
                     }
                 },
                 "example": {
@@ -1769,16 +1757,6 @@
                 "example": "AVRO",
                 "x-codegen-package": "io.apicurio.registry.types"
             },
-            "ArtifactState": {
-                "description": "Describes the state of an artifact or artifact version.  The following states\nare possible:\n\n* ENABLED\n* DISABLED\n* DEPRECATED\n",
-                "enum": [
-                    "ENABLED",
-                    "DISABLED",
-                    "DEPRECATED",
-                    "DELETED"
-                ],
-                "type": "string"
-            },
             "UpdateState": {
                 "title": "Root Type for UpdateState",
                 "description": "",
@@ -1794,6 +1772,17 @@
                 "example": {
                     "state": "DISABLED"
                 }
+            },
+            "ArtifactState": {
+                "description": "Describes the state of an artifact or artifact version.  The following states\nare possible:\n\n* ENABLED\n* DISABLED\n* DEPRECATED\n",
+                "enum": [
+                    "ENABLED",
+                    "DISABLED",
+                    "DEPRECATED",
+                    "DELETED"
+                ],
+                "type": "string",
+                "x-codegen-package": "io.apicurio.registry.types"
             }
         },
         "responses": {
